requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mariadb
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
httpx
websockets
PyQt5
PyQtWebEngine
pandas
openpyxl
Pillow
